+++ Sources/VCore/Views/SwiftUI/Base Button (SwiftUI) (Gesture-Based)/SwiftUIGestureBaseButtonUIViewRepresentable.swift
@@ -1,22 +1,22 @@
 //
-//  SwiftUIBaseButtonNSViewRepresentable.swift
+//  SwiftUIGestureBaseButtonUIViewRepresentable.swift
 //  VCore
 //
-//  Created by Vakhtang Kontridze on 23.03.23.
+//  Created by Vakhtang Kontridze on 12/26/21.
 //
 
-#if canImport(AppKit)
+#if canImport(UIKit) && !os(watchOS)
 
 import SwiftUI
 
-// MARK: - Swift UI Base Button NS View Representable
+// MARK: - SwiftUI Gesture Base Button UI View Representable
 @available(tvOS, unavailable)
-struct SwiftUIBaseButtonNSViewRepresentable: NSViewRepresentable {
+struct SwiftUIGestureBaseButtonUIViewRepresentable: UIViewRepresentable {
     // MARK: Properties
     private let isEnabled: Bool
     private let stateChangeHandler: (BaseButtonGestureState) -> Void
     
-    @State private var gestureRecognizer: AppKitBaseButtonGestureRecognizer?
+    @State private var gestureRecognizer: UIKitBaseButtonGestureRecognizer?
     
     // MARK: Initializers
     init(
@@ -28,22 +28,20 @@ struct SwiftUIBaseButtonNSViewRepresentable: NSViewRepresentable {
     }
 
     // MARK: Representable
-    func makeNSView(context: Context) -> NSView {
-        let view: NSView = .init(frame: .zero)
-
+    func makeUIView(context: Context) -> UIView {
+        let view: UIView = .init(frame: .zero)
+        
         DispatchQueue.main.async(execute: {
-            let gestureRecognizer: AppKitBaseButtonGestureRecognizer = .init(onStateChange: stateChangeHandler)
+            let gestureRecognizer: UIKitBaseButtonGestureRecognizer = .init(onStateChange: stateChangeHandler)
             self.gestureRecognizer = gestureRecognizer
             
             view.addGestureRecognizer(gestureRecognizer)
         })
-
-        //setBindedValues(view, context: context)
-
+        
         return view
     }
-    
-    func updateNSView(_ nsView: NSView, context: Context) {
+
+    func updateUIView(_ uiView: UIView, context: Context) {
         setBindedValues()
     }